+++ src/oci/exceptions.py
@@ -1,6 +1,8 @@
 # coding: utf-8
 # Copyright (c) 2016, 2018, Oracle and/or its affiliates. All rights reserved.
 
+from oci._vendor.requests.exceptions import RequestException as BaseRequestException
+
 
 class ServiceError(Exception):
     """The service returned an error response."""
@@ -109,3 +111,7 @@ class CompositeOperationError(Exception):
     def __init__(self, partial_results=[], cause=None):
         self.partial_results = partial_results
         self.cause = cause
+
+
+class RequestException(BaseRequestException):
+    """An exception occurred when making the request"""