+++ tests/integ/test_retry_service_calls.py
@@ -79,5 +79,5 @@ def test_method_level_retry_strategy_overrides_client_retry_strategy(config):
     client = oci.identity.IdentityClient(config, retry_strategy=ThrowRetryStrategy())
     client.base_client.endpoint = 'https://fakeendpoint.oracle'
 
-    with pytest.raises(oci._vendor.requests.exceptions.ConnectionError):
+    with pytest.raises(oci.exceptions.RequestException):
         client.list_users(config['tenancy'], retry_strategy=oci.retry.NoneRetryStrategy())