+++ tests/unit/test_response_exception.py
@@ -1,0 +1,8 @@
+import oci
+import pytest
+
+
+def test_connection_exception(identity, config):
+    identity.base_client.endpoint = 'https://fakeendpoint.oracle'
+    with pytest.raises(oci.exceptions.RequestException):
+        identity.list_users(config['tenancy'])